requirements_minimal.txt
# Core dependencies for legal document processing pipeline
beautifulsoup4==4.13.4
bs4==0.0.2
langchain==0.2.6
langchain-text-splitters==0.2.2
numpy==2.0.2
python-dotenv==1.0.0
qdrant-client==1.15.0
requests==2.32.4
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.0.0
torch==2.7.1
transformers==4.54.1

# Optional - only needed if using OpenAI in legal_rag_query.py
openai==1.98.0